part2/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
asyncpg>=0.29.0
pandas>=2.1.3
openpyxl>=3.1.2
python-multipart>=0.0.6
google-generativeai>=0.3.2
langchain-google-genai>=0.0.4
langgraph>=0.0.26
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pinecone-client>=3.0.3
langchain>=0.0.347
langchain-community>=0.0.20
langchain-google-genai>=0.0.4
langchain-pinecone>=0.0.2
langgraph>=0.0.26
sqlalchemy>=2.0.23
favicon==0.7.0